proactive/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-socketio[asyncio_client]==5.11.2
pymongo==4.7.2
redis==5.0.4
python-dotenv==1.0.0
